+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+- For `server.allowedHosts` configuration property, add support for matching all subdomains of a host, or all hosts
+
 ## [2.0.0] - August 27, 2020
 - Add SDK key validation configuration
 - Reject request with invalid host (excluding port)